feat(wasm): implement the memory64 proposal
A memory or table may now declare an i64 index type instead of i32,
lifting the four-gibibyte ceiling. Decode, validate, both engines and the
public api surface follow the index type of the memory or table an
instruction names: addresses, sizes and lengths become i64, as do the
results of memory.size, memory.grow, table.size and table.grow, and a
memarg offset immediate widens from u32 to u64.

The official conformance suite is wired up and passes on both engines.

Declared limits and allocated limits are now separate. The specification
lets a module declare up to 2^48 pages for a memory and 2^64-1 entries
for a table, and requires a module declaring more than any host could
allocate to still be *valid* -- test/core/memory64.wast and
test/core/table.wast both define exactly such a module without
instantiating it. Decoding and validation therefore bound the declared
limits by the specification's own ceiling, and the embedder's ceiling is
applied at instantiation: a minimum over it fails to instantiate, and a
maximum over it stops growth earlier. WithMemory64LimitPages configures
it, defaulting to the same four gibibytes a 32-bit memory gets, so
enabling the feature does not by itself let a module claim more host
memory. The table minimum's own ceiling moved out of the decoder for the
same reason.

The 32-bit paths are left as they were. The interpreter routes a 64-bit
memory's scalar loads and stores to their own operation kinds rather than
adding a branch to the existing ones, the native engine branches on the
index type at lowering time, and validation gates its new checks on one
hoisted feature test, so a module that does not use the proposal reaches
the same code it did before. The one exception is memory.grow and
table.grow, whose trampolines now take and return the count as i64
whatever the index type: that costs two instructions in a block that is
about to allocate a whole memory, which is a better trade than a second
trampoline and execution-context field per instruction.

Measured on an Apple M4 (interleaved runs, benchstat, n=16-20): execution
is unchanged on both engines, as are api.Memory access, memory.grow and
module decoding. Validation is +1.0% and interpreter compilation +1.7%,
the residue of resolving an index type per memory instruction; the
compiler engine's compilation is within noise.

See RATIONALE.md for the bounds-check strategy and the limit split.

Closes #40

+++ internal/integration_test/engine/adhoc_test.go
@@ -906,7 +906,7 @@ func testMemOps(t *testing.T, r wazy.Runtime) {
 	require.Zero(t, memory.Size())
 
 	// Any offset should be out of bounds error even when it is less than memory capacity(=memoryCapacityPages).
-	_, err = storeFn.Call(testCtx, wasm.MemoryPagesToBytesNum(memoryCapacityPages)-8)
+	_, err = storeFn.Call(testCtx, wasm.MemoryPagesToBytesNum(uint64(memoryCapacityPages))-8)
 	require.Error(t, err) // Out of bounds error.
 
 	// Try to grow the memory by one page
@@ -915,7 +915,7 @@ func testMemOps(t *testing.T, r wazy.Runtime) {
 	require.Zero(t, results[0]) // should succeed and return the old size in pages.
 
 	// Any offset larger than the current size should be out of bounds error even when it is less than memory capacity.
-	_, err = storeFn.Call(testCtx, wasm.MemoryPagesToBytesNum(memoryCapacityPages)-8)
+	_, err = storeFn.Call(testCtx, wasm.MemoryPagesToBytesNum(uint64(memoryCapacityPages))-8)
 	require.Error(t, err) // Out of bounds error.
 
 	// Check the size command works!
@@ -935,7 +935,7 @@ func testMemOps(t *testing.T, r wazy.Runtime) {
 	require.Equal(t, uint64(memoryCapacityPages), results[0])
 
 	// Now the store instruction at the memory capcity bound should succeed.
-	_, err = storeFn.Call(testCtx, wasm.MemoryPagesToBytesNum(memoryCapacityPages)-8) // i64.store needs 8 bytes from offset.
+	_, err = storeFn.Call(testCtx, wasm.MemoryPagesToBytesNum(uint64(memoryCapacityPages))-8) // i64.store needs 8 bytes from offset.
 	require.NoError(t, err)
 }
